// types/Monitoring.ts
export type Monitoring = {
  cpu_core: number;
  cpu_list: any[];
  cpu_rata: string;
  cpu_type: string;
  durasi: string;
  jumlah_data: number;
  user_agent: string;
  keyword: string;
  metode: string;
  pagination: number;
  paket_download: any[];
  paket_internet: string;
  paket_upload: any[];
  ram_list: any[];
  ram_rata: string;
  ram_tersedia: string;
  ram_total: string;
  waktu_list: any[];
};